requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
Flask-MySQLdb==1.0.1
Zenora==0.0.3.post1
urllib3==1.26.15
mysql-connector-python==8.0.33